Modify the location of classes Add ItemService Re-write Order class Rename LogFileComponent to LogFileService + modify the annotation
