--- message-queue/src/main/java/com/message/queue/messagequeue/main/Order.java
@@ -1,6 +1,0 @@
-package com.message.queue.messagequeue.main;
-
-public class Order {
-
-
-}
